Implemented Solution to The Lucky Deuce Casino
This change provides an implementation of the Daily WTF puzzle "The Lucky Deuce Casino", which emulates a roulette game with the following requirements:

The first requirement is pretty normal. Youre supposed to write a module that generates random numbers like a double-zero roulette wheel. They dont specify how the random numbers are generated, but thats basically a single line of code. Super easy.

Its the second requirement that makes you groan with frustration. They dont want the numbers to be really random. In a true random sequence, the write, the same number may appear many times in a row, just do[sic] to random chance. While that is actually random, it doesnt feel random to our players. They want you to track a history of the random numbers generated, and make something that feels random: numbers that have appeared recently are less like to appear. Runs, where the same number appears 3 times in a row, should never happen.

Alright, you say to yourself, I can do that. But you can do one better. In addition to implementing their requirements, you decide to add in your own. Youre going to write in a cheat function- something that lets you either enter in some sort of cheat code, or in some other fashion makes the output of the roulette wheel predictable. Be careful, though, you dont want to get caught! You'll need to get creative to make sure nobody stumbles on your secret- millions of people are going to gamble with this program.

The "cheat" function provided requires knowledge of the last two numbers, and uses the following formula to generate the next number::

    new_number = ((n1 * n2) + 13) % 37

The only caveat is when a sequence is going to be repeated. The generated number is compared to the last 20 results, and if the number has been generated already, then it is incremented by one. This solves a problem with recurring patterns. Also, the initial two numbers for the formula are randomly generated, so different runs will return different sequences.

To enter cheating mode, the code "pelayo" must be entered in the command line. The game can be restored by entering "reset" at the command line. "exit" can be used to stop playing.